fix(calculator): strip non-semver text
This PR adds a fix to strip non-semver text from a string before converting it to a Version object using strict versioning. In other words, we shouldn't be matching versions like `v9`. We should also be able to handle version tags like `libraries-v1.2.3`, which I don't think was previously supported.